make unsaved-changes-warning component reusable for all legal document types and refactor package structure
