+++ src/main/webapp/app/admin/admin.route.ts
@@ -81,11 +81,11 @@ export const adminState: Routes = [
             },
             {
                 path: 'privacy-statement',
-                loadChildren: () => import('./privacy-statement/privacy-statement-update/privacy-statement-update.module').then((module) => module.PrivacyStatementUpdateModule),
+                loadChildren: () => import('./legal/privacy-statement/privacy-statement-update.module').then((module) => module.PrivacyStatementUpdateModule),
             },
             {
                 path: 'imprint',
-                loadChildren: () => import('./imprint/imprint-update/imprint-update.module').then((module) => module.ImprintUpdateModule),
+                loadChildren: () => import('./legal/imprint/imprint-update.module').then((module) => module.ImprintUpdateModule),
             },
             ...organizationMgmtRoute,
             ...userManagementRoute,
