+++ src/main/webapp/app/admin/legal/imprint/imprint-update.module.ts
@@ -1,22 +1,21 @@
 import { NgModule } from '@angular/core';
-
-import { PrivacyStatementUpdateComponent } from 'app/admin/privacy-statement/privacy-statement-update/privacy-statement-update.component';
 import { ArtemisMarkdownEditorModule } from 'app/shared/markdown-editor/markdown-editor.module';
 import { ArtemisModePickerModule } from 'app/exercises/shared/mode-picker/mode-picker.module';
 import { ArtemisSharedComponentModule } from 'app/shared/components/shared-component.module';
 import { ArtemisSharedCommonModule } from 'app/shared/shared-common.module';
-import { PrivacyStatementUpdateRoutingModule } from 'app/admin/privacy-statement/privacy-statement-update/privacy-statement-update-routing.module';
-import { LegalDocumentUnsavedChangesWarningModule } from 'app/admin/privacy-statement/unsaved-changes-warning/legal-document-unsaved-changes-warning.module';
+import { ImprintUpdateComponent } from 'app/admin/legal/imprint/imprint-update.component';
+import { LegalDocumentUnsavedChangesWarningModule } from 'app/admin/legal/unsaved-changes-warning/legal-document-unsaved-changes-warning.module';
+import { ImprintUpdateRoutingModule } from 'app/admin/legal/imprint/imprint-update-routing.module';
 
 @NgModule({
     imports: [
         ArtemisSharedComponentModule,
         ArtemisSharedCommonModule,
-        PrivacyStatementUpdateRoutingModule,
+        ImprintUpdateRoutingModule,
         ArtemisMarkdownEditorModule,
         ArtemisModePickerModule,
         LegalDocumentUnsavedChangesWarningModule,
     ],
-    declarations: [PrivacyStatementUpdateComponent],
+    declarations: [ImprintUpdateComponent],
 })
-export class PrivacyStatementUpdateModule {}
+export class ImprintUpdateModule {}
