+++ src/main/webapp/app/admin/legal/privacy-statement/privacy-statement-update.component.html
