+++ src/main/webapp/app/admin/legal/unsaved-changes-warning/legal-document-unsaved-changes-warning.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
-import { PrivacyStatementUnsavedChangesWarningComponent } from 'app/admin/privacy-statement/unsaved-changes-warning/privacy-statement-unsaved-changes-warning.component';
+import { UnsavedChangesWarningComponent } from 'app/admin/legal/unsaved-changes-warning/unsaved-changes-warning.component';
 import { ArtemisSharedComponentModule } from 'app/shared/components/shared-component.module';
 import { ArtemisSharedCommonModule } from 'app/shared/shared-common.module';
 
 @NgModule({
     imports: [ArtemisSharedComponentModule, ArtemisSharedCommonModule],
-    declarations: [PrivacyStatementUnsavedChangesWarningComponent],
+    declarations: [UnsavedChangesWarningComponent],
 })
 export class LegalDocumentUnsavedChangesWarningModule {}
