+++ src/main/webapp/app/admin/legal/unsaved-changes-warning/unsaved-changes-warning.component.html
@@ -3,7 +3,7 @@
         <button id="close-modal-btn" type="button" class="btn-close" data-dismiss="modal" aria-hidden="true" (click)="continueEditing()"></button>
     </div>
     <div class="modal-body">
-        <h5>{{ textMessage }}</h5>
+        <h5 jhiTranslate="{{ textMessage }}">Do you really want to proceed and discard the unsaved changes?</h5>
         <div>
             <jhi-button id="discard-content-btn" (click)="discardContent()" [title]="'global.generic.yes'" [ngStyle]="{ 'marginRight.px': 10 }">Yes</jhi-button>
             <jhi-button id="continue-editing-btn" (click)="continueEditing()" [title]="'global.generic.no'">No</jhi-button>
