+++ src/main/webapp/app/admin/legal/unsaved-changes-warning/unsaved-changes-warning.component.ts
@@ -3,10 +3,9 @@ import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
 
 @Component({
     selector: 'jhi-unsaved-changes-warning',
-    templateUrl: './privacy-statement-unsaved-changes-warning.component.html',
-    styleUrls: ['./privacy-statement-unsaved-changes-warning.component.scss'],
+    templateUrl: './unsaved-changes-warning.component.html',
 })
-export class PrivacyStatementUnsavedChangesWarningComponent {
+export class UnsavedChangesWarningComponent {
     @Input()
     textMessage: string;
