+++ src/test/javascript/spec/component/admin/unsaved-changes-warning.component.spec.ts
@@ -1,23 +1,23 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { PrivacyStatementUnsavedChangesWarningComponent } from 'app/admin/privacy-statement/unsaved-changes-warning/privacy-statement-unsaved-changes-warning.component';
+import { UnsavedChangesWarningComponent } from 'app/admin/legal/unsaved-changes-warning/unsaved-changes-warning.component';
 import { ButtonComponent } from 'app/shared/components/button.component';
 import { MockComponent } from 'ng-mocks';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
 import { ArtemisTestModule } from '../../test.module';
 
 describe('UnsavedChangesWarningComponent', () => {
-    let component: PrivacyStatementUnsavedChangesWarningComponent;
-    let fixture: ComponentFixture<PrivacyStatementUnsavedChangesWarningComponent>;
+    let component: UnsavedChangesWarningComponent;
+    let fixture: ComponentFixture<UnsavedChangesWarningComponent>;
     let activeModal: NgbActiveModal;
 
     beforeEach(async () => {
         await TestBed.configureTestingModule({
             imports: [ArtemisTestModule],
-            declarations: [PrivacyStatementUnsavedChangesWarningComponent, MockComponent(ButtonComponent)],
+            declarations: [UnsavedChangesWarningComponent, MockComponent(ButtonComponent)],
         }).compileComponents();
 
-        fixture = TestBed.createComponent(PrivacyStatementUnsavedChangesWarningComponent);
+        fixture = TestBed.createComponent(UnsavedChangesWarningComponent);
         component = fixture.componentInstance;
         activeModal = TestBed.inject(NgbActiveModal);
         fixture.detectChanges();
